Userstory 4 complete
It is possible to search for content by a specific user. If this user
appears as either poster or editor of a post, the post will show up in
the results. The "adressbar" also looks realistic, and changes depending
on what content you are watching, and if you have made a search.