Userstory 10 finished + fixed issues in the database
Methods and small additions to many classes has been made, such as a
method in Content.java to update the "assignto"-field in the post-table
(updates it to the username of the currently logged in user), extensive
fixing in Home.java to make it work as intended with the new
functionality, and the way the labels are added to the contentPane in
the "submitted content"-view has been reworked. Neccessary methods in
Post.java has been added to take into account the "assignedto"-field
added to the post-table in the database. I addition, the column
"assignedto" in the post-table has been changed to default null instead
of not null, as this caused problems trying to get old entries and post
new entries from/to the database.