Added functionality to mark as complete/incomplete.
Added functionality to Home.showHome that makes it possible to (while
viewing submitted content) mark a post as complete. When a user with
copy editor rights (or higher accesslevel) clicks on a post to open the
editor, you can mark it as complete in a checkbox (if it hasnt been
edited before, the value in the database will be null, but once a post
has been edited, it will never be null again). If not checked, the post
(complete-column) will be changed to incomplete (unchanged if it
previously was incomplete).
These changes makes Userstory 22 complete.